Extract widgets related to image annotations
